Add pthreads-based locks
This adds support for multiprocess-safe locks which can be passed by
address. pthreads_mutex was chosen over sem_* because we want a mutex
and pthreads can provide better error handling given this information. I
believe it is also necessary to use the locking primitive exposed by the
threading library Python uses. On Linux, this is pthreads.

The Buffer class supports management of the backing memory. Child
classes must have state which is entirely represented by a buffer of
memory. New buffer objects can be created with the same memory as
another buffer by specifying init=False. Mutating the second buffer will
modify the underlying memory and be visible to the first buffer (with
appropriate memory barriers). It is important that initialization does
not occur unless init=True (the default), since reinitializing an in-use
mutex can cause a deadlock.

Signed-off-by: Sean Anderson <seanga2@gmail.com>