Add atomics
This adds support for C11-based atomics. We supply 32- and 64-bit,
(un)signed integers and doubles. We only use u64 and double, but the
rest are provided for completeness, and to allow for future use. Since
most of the code is shared, the implementation is done by repeatedly
including atomic.h. This is a bit ugly, since we have to paste
identifiers a lot. This could likely be done more easily with C++
templates, but that is left for future improvements.

We use TSO everywhere (the default for non-explicit atomic operations)
to ensure that we use barriers as appropriate. While not every atomic
load/store requires a barrier, enough do that it is simpler to just
default to TSO. This is an area where we could add some memory order
constants in the future.

Signed-off-by: Sean Anderson <seanga2@gmail.com>