Add heap
This adds support for a heap which allocates from shared memory. We
follow a very simple sbrk-style allocation strategy. Reclaiming free'd
memory is not supported. We can't map the whole file up front, so
whenever we dereference a block we map the memory it occupies (if it is
not already mapped). The OS chooses where memory is mapped, so we cannot
assume that addresses will be the same in different processes.
Additionally, we cannot guarantee that memory in memory mapped at
different times will be contiguous. This limits us to allocations of at
most map_size.

To make working with boxed resources easier, blocks returned by malloc
contain a reference to the heap they were allocated from. This requires
that the whole heap be picklable. To implement this, we use a named
temporary file so we can reconstruct the heap from just the name of the
file. However, we want to ensure that we don't have multiple heap
objects with the same backing file in the same process, since they will
not share maps. Because there is a limit on the number of memory maps
available to a process, this is undesirable. To address this, we cache
heaps and only create a new heap if there is not one already created for
this file. This is a bit tricky because we need to keep the lock across
either __new__ and __init__.

We allocate cacheline-aligned by default because most of our allocations
are going to be locks or atomic variables, and this will help reduce
contention.

Signed-off-by: Sean Anderson <seanga2@gmail.com>