Add pickleable generics
Dynamically-created classes in Python cannot normally be pickled, since
Python doesn't know how to create them when unpicking the object. But
since pickle protocol version 4, Python will now walk dotted names by
repeatedly calling getattr. We can (ab)use this by using these paths to
encode arguments to our dynamic classes. This is somewhat like currying
a function in its implementation, especially because these "generics"
can be chained.

This method has a number of limitations. First, the unpicking logic
looks up attributes by splitting on '.'. This means that we cannot use
'.' in attribute names and must replace it or escape it as appropriate.
Second, we do not know a priori when we are done looking up an object.
For fixed-length generics, such as IntType and ProductType, this is not
an issue. However, for ObjectType, we use a sentinel value to indicate
that we are done traversing the namespace tree, and that we should
return the class. To handle nesting (such as Box<Array<Foo, 5>>), we
also include a starting sentinel. < and > are chose for aesthetic
reasons.

To assist in accessing the appropriate object (which could be nested
quite deep), we use the __getitem__ method. This was chosen for its
similarity to the syntax used for type hints. However, it might be
prudent to switch to __call__ to support optional arguments.

Signed-off-by: Sean Anderson <seanga2@gmail.com>