Add metrics
This adds multiprocess-safe metrics following the prometheus_client API.
To avoid the contortions prometheus_client goes through in order to
support both Foo() and Foo().labels(), we use a factory class to select
the appropriate collector. The collector handles creating the metric as
appropriate. It also handles collection, so the metric does not need to
know its name, docstring, or labels. This is done by calling _sample()
in collect() with an add_sample visitor.

LabeledCollector uses a multiprocessing dict to share metrics between
processes. multiprocessing.Server is exceedingly slow, so we use a
per-process cache to speed things up. This means we cannot support
removing metrics, but this is not too bad of a loss because we cannot
free shared memory either. We are not tied deeply to multiprocessing, so
it will be possible to switch to a more performant implementation the
the future.

For simple metrics we can generally use atomic primitives directly.
However, for more complicated metrics, such as Summaries, we cannot do
this because there is more that 64 bits of data to update at once. To
get around this we use the technique described in [1], which appears to
be a varient of seqlocks. Although we technically have a compare/swap
for double values, the observe operation is generally wait free, with
most of the complexity taking place on the _sample side.

[1] https://grafana.com/blog/2020/01/08/lock-free-observations-for-prometheus-histograms/

Signed-off-by: Sean Anderson <seanga2@gmail.com>