metrics: Fix Gauge being unpickleable
Without a wrapper such as Box, we can't use the same name for the gauge
class as the gauge factory. Rename it. Add a test to ensure that all
metrics can be pickled. Do it with a LabeledCollector since tehn we get
a raw metric instead of a Collector (which doesn't need to be
pickleable).

Fixes: e45a08c ("Add metrics")
Signed-off-by: Sean Anderson <seanga2@gmail.com>