types: Rework state-setting functions
__setstate__ doesn't really implement the copy protocol. What we really
want is to be able to supply a block of memory and tell an object
"initialize yourself". Plus, keyword arguments are nice. We use one to
supply the heap, which will be used by Dict.

This generally shifts more information from the pickled part to the part
being pickled. Since we can't free data we should generally try to store
as much as we can in fixed-size parts of the code.

Signed-off-by: Sean Anderson <seanga2@gmail.com>