types: Add shared-memory dict
This is perhaps the simplest shared memory dict there is. We just do a
read-modify-write every time we need to perform an operation. This
should be serialized using a (RW)Lock. This dict frees memory, but our
heap does not, so it effectively takes O(log(n)) space instead of O(n).
We're also limited to a dict of at most map_size (64k) which is around
1900 entries or so (at ~33 bytes per Box plus some initial overhead).

Signed-off-by: Sean Anderson <seanga2@gmail.com>