Remove restriction on heap block size
We do not support blocks larger than the heap map size. This is because
we could have a situation like

	heap = Heap(map_size=4096)
	block1 = heap.malloc(6000)
	block2 = heap.malloc(1)
	block2.deref()
	# heap._maps = [None, MMap(4096, 8192)]
	block1.deref()
	# heap._maps = [MMap(0, 4096), MMap(4096, 8192)]
	# and now block1 isn't continuous!

My solution for this is to align the block size to the map size if it is
larger than the map size. This is a bit wasteful, but no more than our
existing allocator (which never frees anything and fragments like
there's no tomorrow).

Signed-off-by: Sean Anderson <seanga2@gmail.com>