metrics: Add exemplars
This adds exemplars for counters and histograms. In both cases, we use
locks (could probably be done locklessly with RCUs, but that's way too
heavyweight for this). The counter implementation is not terribly
space-efficient, but I don't see this being a big problem as long as you
don't have millions of counters. For histograms, we keep all the
exemplars in one pickled list, as this should help memoize labels.

Signed-off-by: Sean Anderson <seanga2@gmail.com>