heap: Use DupFd to share heaps
Currently, we use the name of the heap to reopen the backing file when
unpickling. However, sometimes the backing file is never deleted, such
as when __del__ is never called. Instead, use an anonymous backing file
which is deleted automatically when all processes exit. Since the file
no longer has a name, we pass its file descriptor using DupFd.

Update the tests to exercise more unusual ways of sharing heaps between
processes. In particular, make sure that we can share a heap with a
third process after the process which created the heap has exited
(test_chain).

Signed-off-by: Sean Anderson <seanga2@gmail.com>