types: Use a persistent id to unpickle the heap
Currently we may store a pickled heap within the heap itself. This is
inefficient, since when unpickling we will always have a copy of the
heap anyway. Use a pesistent id to refer to the heap instead. This
reduces the size of the pickled data. Additionally, we now only need to
unpickle the heap once per process.

Signed-off-by: Sean Anderson <seanga2@gmail.com>